CustomUser model created, profiles model changed, related changes are made
